moved scss file to AwsGraphContainer and DashboardContainer
