+++ app/App.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import Splash from './components/Splash/Splash';
-import DashboardContainer from './containers/DashboardContainer';
+import DashboardContainer from './containers/DashboardContainer/DashboardContainer';
 import './stylesheets/scrollBar.scss';
